Update check_sophos_xg_memory.pl with snmp v3
I wanted to work with this memory check file and i find it useful, thanks, but i needed to work with snmpv3. So i had to add more options for the version 3 of snmp.
User will have just to uncomment the version 3 bloc in line 126 and comment the version 2 in line 138.