Update check_sophos_xg_memory.pl to extend the usage of snmp v3
I wanted to work with the memory check file and i find it useful, thanks, but i needed to work with snmpv3. So i had to add more options for the version 3 of snmp.
User will have just to uncomment the opening session of version 3 bloc in line 126 and comment the version 2 in line 139.